gcp/requirements.txt
tensorflow>=2.17.1,<2.21
google-cloud-storage>=2.10.0
Pillow>=9.0.0
